V updates and prod fixes

+++ time/time.v
@@ -63,12 +63,12 @@ pub fn fps() u32 { return time.fps }
 pub fn now() u64 { return SDL_GetPerformanceCounter() }
 
 // returns the time in milliseconds since the last call
-pub fn laptime(last_time &u64) u64 {
+pub fn laptime(last_time &u64) f64 {
 	mut tmp := last_time
-	mut dt := u64(0)
+	mut dt := f64(0)
 	now := now()
 	if *tmp != 0 {
-		dt = ((now - *tmp) * 1000) / C.SDL_GetPerformanceFrequency()
+		dt = f64((now - *tmp) * 1000) / f64(C.SDL_GetPerformanceFrequency())
 	}
 	*tmp = now
 	return dt
